+++ Examples/UIColor.swift
@@ -1,0 +1,36 @@
+//
+//  UIColor.swift
+//  SwiftCharts
+//
+//  Created by Ivan Schuetz on 17/02/2017.
+//  Copyright © 2017 ivanschuetz. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+
+    public convenience init(hexString: String) {
+        let hexString = hexString.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines) as String
+        let scanner = Scanner(string: hexString)
+        
+        if (hexString.hasPrefix("#")) {
+            scanner.scanLocation = 1
+        }
+        
+        var color: UInt32 = 0
+        scanner.scanHexInt32(&color)
+        
+        let mask = 0x000000FF
+        let r = Int(color >> 16) & mask
+        let g = Int(color >> 8) & mask
+        let b = Int(color) & mask
+        
+        let red   = CGFloat(r) / 255.0
+        let green = CGFloat(g) / 255.0
+        let blue  = CGFloat(b) / 255.0
+        
+        self.init(red:red, green:green, blue:blue, alpha:1)
+    }
+    
+}