+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add custom transtorm mode to ChartPointsViewsLayer
 - Allow to disable axis layer modification of chart's frame
 - Allow to show chart points outside of the chart's inner frame 
+- Add example for ranged axis and rotation handling
 
 ## [0.5] - 2016.11.23
 - Add multi-chart tracker layer